refactor: move webpack config into it's own folder in config
