+++ package.json
@@ -4,7 +4,7 @@
   "private": true,
   "scripts": {
     "analyze-bundle": "npm run build && npx source-map-explorer 'build/static/js/*.js'",
-    "start": "node scripts/start.js",
+    "dev": "node scripts/dev.js",
     "build": "node scripts/build.js",
     "test": "jest --coverage",
     "test:watch": "jest --watch",
